// src/app/models/property.model.ts
export class Property {
    _id: string;
    name: string;

    address: string;
    city: string;
    state: string;
    pinCode: number;

    phone: number;
    email: string;

    numberOfRooms: number;
    availableRooms: number;
    tarrif: number;
    images: string[];
    createdOn: string;
    lastUpdate: string;
}